+++ HISTORY.rst
@@ -64,9 +64,8 @@ Release History
 2.0.1 (2015-05-27)
 ------------------
 
-* Handle new environment variables
+* Handle new environment variable
   `PLATFORMIO_BUILD_FLAGS <http://docs.platformio.org/en/latest/envvars.html#platformio-build-flags>`_
-  and `PLATFORMIO_LDF_CYCLIC <http://docs.platformio.org/en/latest/envvars.html#platformio-ldf-cyclic>`_
 * Pass to API requests information about Continuous Integration system. This
   information will be used by PlatformIO-API.
 * Use ``include`` directories from toolchain when initialising project for IDE